@@ -1582,7 +1582,7 @@ bool AppInitMain(InitInterfaces& interfaces)
             }
                 auto tip = chainActive.Tip();
                 assert(tip);
-                if (!CClaimTrieCache(pclaimTrie).validateDb(tip->hashClaimTrie))
+                if (!CClaimTrieCache(pclaimTrie).validateDb(tip->hashClaimTrie)) {
                     strLoadError = _("Error loading the claim trie from disk");
                     break;
                 }
